Cassandra csv generation for small n entries implemented
